スロットidのintをshortに

+++ Assets/Scripts/MainGame/Network.Interface/Send/ISendPlaceHotBarBlockProtocol.cs
@@ -2,6 +2,6 @@
 {
     public interface ISendPlaceHotBarBlockProtocol
     {
-        public void Send(int x, int y, int hotBarSlot,int playerId);
+        public void Send(int x, int y, short hotBarSlot,int playerId);
     }
 }
